refactor: change name of typeAlises.ts file to global.ts
