+++ src/types/global.ts
@@ -1,4 +1,3 @@
 export type Variant = 'default' | 'withDisable' | 'login' | 'placeholder' | 'animate' | 'search';
 export type InputType = 'text' | 'password' | 'email' | 'search' | 'textarea';
-export type Size = 'verysmall' | 'small' | 'medium' | 'large';
 export type ButtonType = 'button' | 'submit' | 'reset';
